Decoding is up.
LSB - not done
Result page- not done
For encoding, path is testing/
For decoding, path is decode/
Choose any images to encode, Once encode is done, it is stored inside images -> input -> stego images.
The image inside input should be the original photo. ( this is to link up with the result webpage that isnt done yet.